uint256: Correct base 10 output formatting.
The corrects a formatting issue where the base 10 output is not always
printing the expected results.

It also adds a test which triggers the condition as well as a new
randomized test to compare the output of random values against the big
int output for all of the supported bases.