chaincfg: Introduce subsidy split change agenda.
This adds a new definition for the upcoming agenda vote to change the
subsidy split to 10% PoW, 80% PoS, and 10% Treasury defined in DCP0010
along with two new parameters for the modified values.

It does not include any code to make decisions based on the status of
the agenda or bump block versions.  It only adds the definition for the
agenda.