rpcclient: Prepare v7.0.0.
This updates the rpcclient module dependencies, the copyright year in
the files modified since the previous release, and serves as a base for
rpcclient/v7.0.0.

The updated direct dependencies in this commit are as follows:

- github.com/decred/dcrd/gcs/v3@v3.0.0
- github.com/decred/dcrd/rpc/jsonrpc/types/v3@v3.0.0

The full list of updated direct dependencies since the previous
rpcclient/v6.0.2 release are as follows:

- github.com/decred/dcrd/chaincfg/chainhash@v1.0.3
- github.com/decred/dcrd/dcrjson/v4@v4.0.0
- github.com/decred/dcrd/dcrutil/v4@v4.0.0
- github.com/decred/dcrd/gcs/v3@v3.0.0
- github.com/decred/dcrd/rpc/jsonrpc/types/v3@v3.0.0
- github.com/decred/dcrd/txscript/v4@v4.0.0
- github.com/decred/dcrd/wire@v1.5.0
- github.com/decred/slog@v1.2.0

Finally, all modules in the repository that depend on the module are
tidied to ensure they are updated to use the latest versions hoisted
forward as a result.