dcrutil: Limit WIF decode to max possible size.
This modifies DecodeWIF to reject any attempts to decode strings that
are larger than the max possible size early since there is no reason to
waste time and memory doing the base 58 decode when it is guaranteed to
be invalid anyway.

It also adds an associated test to ensure proper functionality.