rpcserver: Update websocket ping timeout handling.
Currently, timeouts when responding to a websocket ping control message
with a pong message cause the websocket client to be disconnected, but
these types of failures are nearly always temporary network errors due
to things such as congestion and thus should not result in
disconnection.

In order to make that the case, this modifies the websocket ping handler
to ignore timeout network errors when attempting to respond with a pong
and also takes this opportunity to increase the timeout to 5 seconds
from 1 second to support multi-continent communication.