edwards: More strict pubkey parsing.
This modifies the public key parsing method to make it more strict by
rejecting public keys that are not exactly the required length as
opposed to silently truncating or padding them and adds tests
accordingly.

This helps prevent misuse by callers and makes it behave more
consistently with the secp256k1 public key parsing.

Note that this does not affect consensus code since the only place it is
called is during the signature verification opcode where the consensus
rules ensure the the key is the correct length before parsing it.