ffldb: Add dbErr to error description.
Add the reason for bad inserts, updates, and deletes to the error
description and display key as hex in order to give the user more
information when the database errors. This is especially useful for
upgrades which may fail for several reasons, such as low disc space,
which the user will now be informed about.