secp256k1: Support go generate w/o removing file.
This modifies the logic that deals with generating, loading, and
decompressing the compressed byte points to allow go generate to run
correctly regardless of whether or not the generated file already
exists.

It accomplishes this by making use of a function to obtain the
compressed byte points string instead of directly referencing the
variable such that the function is nil by default and only set to the
real function via the generated file.