stake: Tx ver as agenda proxy in DetermineTxType.
This updates stake.DetermineTxType to use the tx version to check if the
treasury agenda is active rather than relying on the agenda flag.

This is safe since the treasury agenda has activated and there are no
transactions with a transaction version greater than or equal to the
treasury transaction version in any blocks prior to the agenda
activating.