server: Remove unnecessary vote ntfn check.
This removes the check for whether or not the block height is after the
height of the final checkpoint when deciding whether to send a vote
notification since it is no longer necessary or has any effect.

The reason it no longer has any effect is because there is a check prior
to it that ensures the the chain is considered current and the more
recent headers-first syncing means the chain will not be considered
current until the block associated with the best known header is reached
which is always ahead of any checkpoint.