blockchain: Fix revocation fee limit bug.
This retroactively fixes a bug that resulted in the fee limit not being
properly enforced for revocation transactions.  The exploitation risk of
this was that in a split transaction, one party could intentionally
increase the fee paid to miners for another party.

It is no longer possible to exploit this on mainnet since the automatic
revocations agenda has activated, which enforces that the fee must be
zero for revocation transactions.

The fee limit for revocation transactions was never violated on mainnet
so this logic can be safely corrected without impacting consensus.  This
has been validated by running a full sync with checkpoints and assume
valid disabled.