mempool: Explicitly reject standalone treasurybase.
This updates the mempool to explicitly reject standalone treasurybases
and adds a test to ensure proper functionality.

Note that treasurybases are already rejected because IsCoinBaseTx
currently detects a treasurybase as a coinbase as well, but it is safer
to be explicit in case the coinbase detection function is updated in the
future to exclude treasurybases.

Also since treasurybases are rejected early, remove the code that deals
with special casing them later given there is no way the transaction is
one.