cpuminer: Rework speed stat tracking.
This reworks the speed stat tracking to make it more accurate as well as
significantly reduce lock contention.

A high level overview of the changes is as follows:

- Each worker is given its own stats state to independently update
  - Each worker now has a unique ID which is used to store and clean up
    said state
- The elapsed hashing time per worker is now tracked
- The individual speed stats are updated via lock-free atomic primitives
  versus the previous single mutex that all workers shared
- Each worker now periodically updates its individual speed stats every
  fixed number of nonces instead of relying on a ticker
- The speed monitor goroutine periodically sums the results from all of
  the individual worker stats

Some local test mining shows that prior to these changes the reported
hash speed varies pretty wildly between 43 kh/s and 427 kh/s and more
wild swings with more workers.  With the changes, the reporting is more
in line with what I would expect without additional changes and comes in
a much tigher (and a bit higher due to a bit of reduced lock contention)
range of 408 kh/s and 568 kh/s.