[release-v1.7] mining: Copy regular txns for alternate templates.
This modifies the template generation code in the case an alterative
block building on the parent of the current tip is needed due to lack of
votes to also copy the regular transactions from the current tip in
addition to the stake transactions.

This is useful because it helps make it the template creation logic more
robust in the case of duplicate spends and also helps ensure all of the
same transactions, such as split transactions make it into the new
alternative block templates.