[release-v1.7] rpcserver: Return template errors from getwork RPC.
This modifies the RPC server getwork method to return an error in the
case of a failed background template.  Callers generally expect an error
as opposed to long blocking behavior in the case there is a temporary
reason block templates are unable to be generated.

Also, correct the offsets in the comments regarding the serialized
locations of the fields while here.