blockchain: Make longer running tests parallel.
This makes most of the tests that take a while to run parallel to
significantly reduce the overall testing time roughly in half.

Note that this does cause the output of the various tests that log the
creation of the blocks and their acceptance/rejection to be commingled,
but the testing framework adds a header for which test each line comes
from, so it's fairly easy to reconstruct the logging stream for any
individual test in the event of any test failures.