indexers: Remove unused PrevScripts from idx ntfn.
This removes the no longer used PrevScripts field from the index
notification.

This is part of an overall effort to remove all of the code related to
taking previous script snapshots and feeding them through to the
indexers since there are no longer any indexers that need access to it.