blockchain: Remove unused PrevScripts from ntfns.
This removes the no longer used PrevScripts field from the
NTBlockConnected and NTBlockDisconnected notifications as well as the
code that takes a snapshot populates the fields.

This is part of the overall effort to remove all of the code related to
taking previous script snapshots and feeding them through to the
indexers.