indexers: Remove unused ChainQueryer.PrevScripts.
This removes the no longer used PrevScripts method from the ChainQueryer
interface as well as the methods that implemented it from the tests and
the blockchain query adapator.

This is part of the overall effort to remove all of the code related to
taking previous script snapshots and feeding them through to the
indexers.