blockchain: PrevScripter interface for script val.
This exposes the PrevScripter interface used elsewhere in the code from
blockchain and modifies the exported ValidateTransactionScripts method
and associated internal implementation to accept the interface instead
of a concrete UtxoView.

This helps reduce the scope of script validation dependencies and thus
provides the opportunity to reduce the implementation complexity of
callers.