mempool: Store transaction descs in pools.
Currently, several places that interact with the maps that house the raw
transactions and staged transactions need access to the additional
description data for the transaction in addition to the transaction
itself.  This involves an additional lookup and the need to pass around
the map that houses that associated description data to perform the
lookup.

In order to avoid that additional lookup, passing around the associated
map, and to simplify the code, this modifies those maps to store the
overall transaction descriptions directly instead of the raw
transactions and updates the various consumers accordingly.