main: Tweak runtime GC params for Go 1.19.
Go 1.19 introduced the ability to specify a soft upper memory limit to
help deal with transient memory spikes without needing to set the GC
percent to an artificially low value.

This takes advantage of that to impose an upper memory limit that leaves
plenty of headroom for the minimum recommended value while taking into
account the max utxo cache size configuration option and increase the GC
percent to the default value of 100% which allows the mem usage to
expand more in between each GC cycle which in turn significantly reduces
the number of GC cycles that need to be performed, particularly while
performing the initial chain sync.

The end result is much less CPU time spent doing garbage collection and
thus reduces the amount of time it takes to perform the initial chain
sync by about 10%.

The update also has the nice side benefit that the `GOGC` environment
variable can be used if an advanced sysadmin really wanted to tune it
since it is no longer being overridden.