server: Bump supported wire pver for reject removal.
This bumps the supported wire protocol version for the server to the
version that indicates the reject message has been removed.

This probably should have been done as a part of the v1.7.0 release so
that peers would negotiate this version, but it wasn't, so rejecting
older protocol versions will need to wait until this code is included in
a release that also contains a consensus rule upgrade to piggyback on.