peer: Use latest pver by default.
This updates the max and default protocol version the peer package uses
to the latest protocol version (wire.RemoveRejectVersion) and updates
the tests to account for the removal of the reject message.

It also adds a test that explicitly tests the handler on older protocol
versions since the associated message is no longer sent on the latest
one.